Move `callMiddleware()`.
